Website with frontend for uploading images
Django backend with API for receiving images. Frontend in React.js
