+++ website-building/backend/SPI/models.py
@@ -10,3 +10,11 @@ class Todo(models.Model):
 
     def _str_(self):
         return self.title
+
+class Post(models.Model):
+    title = models.CharField(max_length=100)
+    content = models.TextField()
+    image = models.ImageField(upload_to='post_images')
+    
+    def __str__(self):
+        return self.title
