+++ website-building/backend/SPI/serializers.py
@@ -1,10 +1,15 @@
 # todo/serializers.py
 
 from rest_framework import serializers
-from .models import Todo
+from .models import Todo, Post
 
 
 class TodoSerializer(serializers.ModelSerializer):
     class Meta:
         model = Todo
         fields = ('id', 'title', 'description', 'completed')
+
+class PostSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Post
+        fields = '__all__'
