+++ website-building/backend/SPI/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path('posts/', views.PostView.as_view(), name= 'posts_list'),
+]
