+++ website-building/frontend/.eslintcache
@@ -1,1 +1,1 @@
-[{"/home/dennis/GIT/website-building/frontend/src/index.js":"1","/home/dennis/GIT/website-building/frontend/src/App.js":"2","/home/dennis/GIT/website-building/frontend/src/reportWebVitals.js":"3","/home/dennis/GIT/website-building/frontend/src/components/Modal.js":"4"},{"size":547,"mtime":1608470666345,"results":"5","hashOfConfig":"6"},{"size":3852,"mtime":1608471167611,"results":"7","hashOfConfig":"6"},{"size":362,"mtime":1608468717173,"results":"8","hashOfConfig":"6"},{"size":2015,"mtime":1608470782476,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"16sn34t",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/dennis/GIT/website-building/frontend/src/index.js",[],"/home/dennis/GIT/website-building/frontend/src/App.js",[],"/home/dennis/GIT/website-building/frontend/src/reportWebVitals.js",[],["19","20"],"/home/dennis/GIT/website-building/frontend/src/components/Modal.js",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},"no-native-reassign",["25"],"no-negated-in-lhs",["26"],"no-global-assign","no-unsafe-negation"]
+[{"/home/dennis/GIT/solar-panel-inspection/website-building/frontend/src/index.js":"1","/home/dennis/GIT/solar-panel-inspection/website-building/frontend/src/reportWebVitals.js":"2","/home/dennis/GIT/solar-panel-inspection/website-building/frontend/src/App.js":"3","/home/dennis/GIT/solar-panel-inspection/website-building/frontend/src/components/Modal.js":"4"},{"size":547,"mtime":1608559179876,"results":"5","hashOfConfig":"6"},{"size":362,"mtime":1608559179876,"results":"7","hashOfConfig":"6"},{"size":1766,"mtime":1608563347838,"results":"8","hashOfConfig":"6"},{"size":2015,"mtime":1608559179876,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"1p9kxu6",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/dennis/GIT/solar-panel-inspection/website-building/frontend/src/index.js",[],["19","20"],"/home/dennis/GIT/solar-panel-inspection/website-building/frontend/src/reportWebVitals.js",[],"/home/dennis/GIT/solar-panel-inspection/website-building/frontend/src/App.js",[],"/home/dennis/GIT/solar-panel-inspection/website-building/frontend/src/components/Modal.js",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},"no-native-reassign",["25"],"no-negated-in-lhs",["26"],"no-global-assign","no-unsafe-negation"]
