+++ website-building/frontend/src/App.css
@@ -2,6 +2,21 @@
   text-align: center;
 }
 
+#image {
+  color: white;
+}
+
+#header {
+  color: white;
+  text-align: center;
+  font-size: 30px;
+}
+
+#header-place {
+  margin-top: 3%;
+  margin-bottom: 6%;
+}
+
 .App-logo {
   height: 40vmin;
   pointer-events: none;
